task13-browser-use/pyproject.toml
[project]
name = "task13-browser-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.1.40",
    "fastapi>=0.115.12",
    "langchain-google-genai>=2.1.2",
    "python-dotenv>=1.0.0",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.0",
]
